Fix default reactive handlers to only run once
Reactive has some unexpected 'gotchas' which arise due to the
different concepts involved.  Non reactive charms rely on 'hooks' that
indicate exactly which function will run when a hook is invoked.
Reactive, on the other hand, only uses hooks to run the reactive system,
and, once it is running, ANY handler that has conditions that evaluate
to true will run.

This patch modifies the default handlers that charms.openstack provides
such that they only execute ONCE for each associated reactive state as
set by the associated interface.  They use the 'detaching' state (i.e.
when the state goes away) to reset the 'run_once' guard state.  This
should mean that charms that use the default state will have 'quieter'
debug-logs and, perhaps, more understandable behaviour.

Note that the charms that use charms.openstack also need to have a
similar guard around their own, non default, handlers.

Note the depends-on is simply because this change is linked to the
charm refactor.

Change-Id: I0eb50127aac88a8bf7ee431495a6df0c60ef3fce
Depends-On: Ifd171b5a47b814cee6c7e53d9a6bae52833caed2